move minisocket.js
